# sources/api.yaml
openapi: 3.0.3
info:
    title: ADV
    description: "TEST"
    version: 0.0.1
    contact:
        name: "Mirko Colageo"
        email: m.colageo@almaviva.it
paths:
...
    /feature-importance:
        post:
            summary: Return the feature importance data
            description: Return the feature importance data for a given pilot
            operationId: featureImportance
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/FeatureImportanceRequest"
            responses:
                "200":
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/FeatureImportanceResponse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/FeatureImportanceResponse"
                "400":
                    description: Invalid IDs combination supplied
                "404":
                    description: Pilot / Model not found
                "500":
                    description: Internal server error
            tags:
                - featureImportance
...
servers: []
components:
    securitySchemes:
        bearer:
            scheme: bearer
            bearerFormat: Bearer
            description: "Please enter token in following format: Bearer <JWT>"
            type: http
    schemas:
        FeatureImportanceRequest:
            type: object
            properties:
                model:
                    description: the identifier of the model file name
                    example: "model.pkl"
                    type: string
                meta_data:
                    description: the identifier of the metadata file name
                    example: "metadata.json"
                    type: string
            required:
                - model
        FeatureImportanceResponse:
            type: object
            properties:
                data:
                    type: string
...